+++ services/src/test/java/ru/dpopkov/knownics/dto/converters/QuestionToDtoTest.java
@@ -4,14 +4,17 @@ import org.junit.jupiter.api.Test;
 import ru.dpopkov.knownics.domain.KeyTerm;
 import ru.dpopkov.knownics.domain.Language;
 import ru.dpopkov.knownics.domain.TextType;
-import ru.dpopkov.knownics.domain.answer.Answer;
+import ru.dpopkov.knownics.domain.answer.*;
 import ru.dpopkov.knownics.domain.question.Category;
 import ru.dpopkov.knownics.domain.question.Question;
 import ru.dpopkov.knownics.domain.question.QuestionText;
+import ru.dpopkov.knownics.dto.AnswerDto;
 import ru.dpopkov.knownics.dto.QuestionDto;
 import ru.dpopkov.knownics.dto.TranslationDto;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -19,7 +22,8 @@ class QuestionToDtoTest {
 
     private static final LocalDateTime NOW = LocalDateTime.now();
 
-    final QuestionToDto toDto = new QuestionToDto(new CategoryToDto(), new TranslationToDto(), new KeyTermToDto());
+    final QuestionToDto toDto = new QuestionToDto(new CategoryToDto(), new TranslationToDto(),
+            new AnswerToDto(new TranslationToDto(), new SourceDetailsToDto()), new KeyTermToDto());
 
     @Test
     void testConvert() {
@@ -33,7 +37,12 @@ class QuestionToDtoTest {
         QuestionText questionText = new QuestionText(Language.EN, TextType.PLAINTEXT, "What is Java");
         question.addTranslation(questionText);
         final Answer answer = new Answer();
-        answer.setId(12L);
+        final long answerId = 12L;
+        answer.setId(answerId);
+        answer.setType(AnswerType.ORIGINAL);
+        answer.setSourceDetails(new SourceDetails(new Source("Core Book", SourceType.BOOK), "details"));
+        final String answerText = "What is JUnit";
+        answer.addTranslation(new AnswerText(Language.EN, answerText));
         question.getAnswers().add(answer);
         question.setPreferredAnswer(answer);
         question.getKeyTerms().add(new KeyTerm("Core Java", "Test description"));
@@ -56,8 +65,14 @@ class QuestionToDtoTest {
         assertEquals(questionText1.getType().toString(), questionTextDto.getType());
         assertEquals(questionText1.getText(), questionTextDto.getText());
 
-//        assertEquals(question.getAnswers().size(), dto.getAnswerIds().size());
-//        assertEquals(question.getPreferredAnswer().getId(), dto.getPreferredAnswerId());
+        // Answers
+        assertEquals(1, dto.getAnswers().size());
+        List<AnswerDto> answerDtoList = new ArrayList<>(dto.getAnswers());
+        assertEquals(answerId, answerDtoList.get(0).getId());
+        List<TranslationDto> answerTranslations = new ArrayList<>(answerDtoList.get(0).getTranslations().values());
+        assertEquals("EN", answerTranslations.get(0).getLanguage());
+        assertEquals(answerText, answerTranslations.get(0).getText());
+
         assertEquals(question.getKeyTerms().size(), dto.getKeyTerms().size());
         assertEquals(question.getComment(), dto.getComment());
     }